+++ app/Http/Controllers/Auth/SubscriptionController.php
@@ -201,8 +201,8 @@ class SubscriptionController extends Controller
             'limit' => $limit
         ]), [
             'subscription_id' => ['bail', 'required', 'exists:subscriptions,id'],
-            '$page' => ['bail', 'sometimes', 'numeric'],
-            'limit' => ['bail', 'sometimes', 'numeric'],
+            '$page' => ['bail', 'sometimes', 'integer'],
+            'limit' => ['bail', 'sometimes', 'integer'],
         ]);
 
         if ($validator->fails()) {
